Move markdown publishers to submodule
